#19 changed the data type of the price to reflect a 2 decimal place number

+++ Backend/src/main/java/teamproject/backend/Model/MenuItem.java
@@ -25,7 +25,7 @@ public class MenuItem {
   private int quantity;
 
   @Column (nullable = false, precision =10, scale =2)
-  private float price;
+  private BigDecimal price;
 
   @Column (nullable = false)
   private String imageURL;
